+++ changelog.md
@@ -6,6 +6,16 @@ This changelog is based on [Keep A Changelog](https://keepachangelog.com/en/1.1.
 
 ## Changed
 
+## Removed
+
+## Fixed
+
+## Added
+
+# v1.4.0
+
+## Changed
+
 * `genesis_utxo` and `registration_utxo` no longer have to have no native tokens.
 * Update ogmios to v6.9.0
 * Organized Rust sources into two directories: toolkit and node.
@@ -17,15 +27,13 @@ This changelog is based on [Keep A Changelog](https://keepachangelog.com/en/1.1.
 * * pallets are not generic on SidechainParams anymore, they use UtxoId (genesis_utxo) instead
 * * migration of pallet_sidechain storage is not present in this update, it has to be done before v1.4, so as for now it is a very breaking change
 
-## Removed
-
 ## Fixed
 
 * CardanoNetwork bug in `partner-chains-cli`, that would cause the CLI to fail with the mainnet.
 
 ## Added
 
-* Added `smart-contracts` command to the node with first command `init-governance`.
+* Added `smart-contracts` command to the node with sub-commands `init-governance` and `get-scripts`.
 
 # v1.3.0